@@ -1,16 +1,16 @@
 {
   "name": "task-browser",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "1.2.35",
-    "@fortawesome/free-solid-svg-icons": "5.15.3",
-    "@fortawesome/react-fontawesome": "0.1.14",
-    "axios": "0.21.1",
-    "bootstrap": "5.0.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router-dom": "5.2.0"
+    "@fortawesome/fontawesome-svg-core": "6.1.1",
+    "@fortawesome/free-solid-svg-icons": "6.1.1",
+    "@fortawesome/react-fontawesome": "0.1.18",
+    "axios": "0.27.2",
+    "bootstrap": "5.1.3",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-router-dom": "6.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -34,15 +34,15 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "5.14.1",
-    "@testing-library/react": "11.2.7",
-    "@testing-library/user-event": "13.1.9",
-    "@types/jest": "26.0.23",
-    "@types/node": "12.20.15",
-    "@types/react": "17.0.11",
-    "@types/react-dom": "17.0.8",
-    "@types/react-router-dom": "5.1.5",
-    "react-scripts": "4.0.3",
-    "typescript": "4.3.4"
+    "@testing-library/jest-dom": "5.16.4",
+    "@testing-library/react": "13.2.0",
+    "@testing-library/user-event": "14.2.0",
+    "@types/jest": "27.5.1",
+    "@types/node": "16.11.35",
+    "@types/react": "18.0.9",
+    "@types/react-dom": "18.0.4",
+    "@types/react-router-dom": "5.3.3",
+    "react-scripts": "5.0.1",
+    "typescript": "4.6.4"
   }
 }
